add sanity check to Waydroid launcher
this is helpful after doing a SteamOS update the waydroid binary will be gone. this will display a dialog box if the waydroid binary is missing and tell the user to re-run the install script.

+++ steamos-waydroid-installer.sh
@@ -198,6 +198,17 @@ echo -e "$current_password\n" | sudo -S chown root:root /etc/sudoers.d/zzzzzzzz-
 cat > ~/Android_Waydroid/Android_Waydroid_Cage.sh << EOF
 #!/bin/bash
 
+# check if waydroid exists
+if ! [ -f /usr/bin/waydroid ]
+then
+        kdialog --sorry "Cannot start Waydroid. Waydroid does not exist! \\
+        \\nIf you recently performed a SteamOS update, then you also need to re-install Waydroid! \\
+        \\nLaunch the Waydroid install script again to re-install Waydroid! \\
+        \\nSteamOS version: \$(cat /etc/os-release | grep -i VERSION_ID | cut -d "=" -f 2) \\
+        \\nKernel version: \$(uname -r | cut -d "-" -f 1-5)"
+        exit
+fi
+
 export shortcut=\$1
 
 killall -9 cage &> /dev/null
